+++ src/GameEngine.h
@@ -9,11 +9,11 @@ class Font;
 class GameEngine {
 
 private:
-  static void (*_GameUpdate)(float);
+  static void(*_GameUpdate)(double);
   static void (*_GameRender)();
   static bool run;
 
-  static void update(const float delta);
+  static void update(const double delta);
   static void render();
   static void Postrender();
 
@@ -22,10 +22,10 @@ public:
   static void init();
   static void Loop();
   static void StopLoop();
-  static void RegisterUpdate(void (*GameUpdate)(float));
+  static void RegisterUpdate(void(*GameUpdate)(double));
   static void RegisterRender(void (*GameRender)());
   //
-  static unsigned int getTime();
+  static double getTime();
   //
   static Renderer* Renderer;
   static Meshloader* Meshloader;
